Server Update
Summary:
Created a working lobby system. The desktop "Creates" the lobby and the
mobile users are able to join. After that, the  server handels most of
the communication of said lobby.

Added:
- A lobby class that easly creates and handels most of the dirty work of
the lobby. This way i can easly track what the lobby is doing and it
makes the code cleaner to read and write

- Updated the server script to handel those lobby's in a clean and
understandble way.

- A pincode system where you can join a lobby be entering a pincode.

- A client scripts that asks to create a lobby. After that the server
creates one and gives the pincode the desktop client.

Note:
We need to find out what te limit is of how many lobby's created. It
must not lag or drop the speed of the connection in any way. testing is
required.